Working on having one N for each Column or Table (or node?) in order to reset them when each node begin their execution.
This will allow adding customers at a date and all their orders after that date, and just after that date (not random length into the future)

+++ SQLDataProducer.Tests/NodeIteratorTest.cs
@@ -193,7 +193,7 @@ namespace SQLDataProducer.Tests
 
 
             NodeIterator it = new NodeIterator(customer);
-            int actualExpectedCount = it.GetExpectedInsertCount();
+            long actualExpectedCount = it.GetExpectedInsertCount();
             Assert.That(actualExpectedCount, Is.EqualTo(14));
         }
 
@@ -203,7 +203,7 @@ namespace SQLDataProducer.Tests
         {
             ExecutionNode customer = ExecutionNode.CreateLevelOneNode(2, "No tables");
             NodeIterator it = new NodeIterator(customer);
-            int actualExpectedCount = it.GetExpectedInsertCount();
+            long actualExpectedCount = it.GetExpectedInsertCount();
             Assert.That(actualExpectedCount, Is.EqualTo(0));
 
         }
